+++ supvisors/tests/test_viewapplication.py
@@ -270,11 +270,11 @@ def test_get_process_data(mocker, view):
     """ Test the ViewApplication.get_process_data method. """
     # patch the selected application
     process_1 = Mock(application_name='appli_1', process_name='process_1', namespec='namespec_1',
-                     running_identifiers=set(), state='stopped', rules=Mock(expected_load=20),
+                     running_identifiers=set(), state='stopped', has_crashed=False, rules=Mock(expected_load=20),
                      **{'state_string.return_value': 'stopped'})
     process_2 = Mock(application_name='appli_2', process_name='process_2', namespec='namespec_2',
                      running_identifiers=['10.0.0.1', '10.0.0.3'],  # should be a set but hard to test afterwards
-                     state='running', rules=Mock(expected_load=1),
+                     state='running', has_crashed=True, rules=Mock(expected_load=1),
                      **{'state_string.return_value': 'running'})
     view.application = Mock(processes={process_1.process_name: process_1, process_2.process_name: process_2})
     # patch context
@@ -284,11 +284,11 @@ def test_get_process_data(mocker, view):
     # test call
     data1 = {'application_name': 'appli_1', 'process_name': 'process_1', 'namespec': 'namespec_1',
              'identifier': '10.0.0.1', 'statename': 'stopped', 'statecode': 'stopped', 'gravity': 'stopped',
-             'running_identifiers': [], 'description': 'something',
+             'has_crashed': False, 'running_identifiers': [], 'description': 'something',
              'expected_load': 20, 'nb_cores': 4, 'proc_stats': mocked_stats}
     data2 = {'application_name': 'appli_2', 'process_name': 'process_2', 'namespec': 'namespec_2',
              'identifier': '10.0.0.1', 'statename': 'running', 'statecode': 'running', 'gravity': 'running',
-             'running_identifiers': ['10.0.0.1', '10.0.0.3'], 'description': 'something',
+             'has_crashed': True, 'running_identifiers': ['10.0.0.1', '10.0.0.3'], 'description': 'something',
              'expected_load': 1, 'nb_cores': 4, 'proc_stats': mocked_stats}
     assert view.get_process_data() == [data1, data2]